@@ -14,6 +14,7 @@
     "@stomp/stompjs": "^7.0.0",
     "axios": "^1.6.7",
     "bootstrap": "^5.3.2",
+    "bootstrap-vue": "^2.23.1",
     "chart.js": "^4.4.1",
     "firebase": "^10.8.0",
     "firebase-tools": "^13.1.0",
@@ -27,7 +28,7 @@
     "socket.io-client": "^4.7.4",
     "v-calendar": "^3.1.2",
     "vite-plugin-pwa": "^0.17.5",
-    "vue": "^3.3.11",
+    "vue": "^3.4.19",
     "vue-chartjs": "^5.3.0",
     "vue-chartkick": "^1.1.0",
     "vue-color": "^2.8.1",
@@ -38,6 +39,7 @@
     "vue3-cookies": "^1.0.6",
     "vuefire": "^3.1.22",
     "vuex": "^4.1.0",
+    "web-audio-recorder": "^1.0.5",
     "web-audio-recorder-js": "^0.0.2",
     "web-push": "^3.6.7",
     "webrtc-adapter": "^8.2.3",
