@@ -1,7 +1,6 @@
 // stores/todoStore.js
 import { defineStore } from 'pinia';
-import { getTodoListByDate } from '@/api/todos'; // API 호출을 위한 함수
-
+import { getTodoList } from '@/api/todos'; // API 호출을 위한 함수
 export const useTodoStore = defineStore('todo', {
   state: () => ({
     todos: [], // Todo 항목들을 저장할 배열
@@ -9,7 +8,7 @@ export const useTodoStore = defineStore('todo', {
   actions: {
     async fetchTodos(date) {
       try {
-        const response = await getTodoListByDate(date); // API 호출
+        const response = await getTodoList(date); // API 호출
         this.todos = response; // 응답으로 받은 Todo 목록으로 상태 업데이트
       } catch (error) {
         console.error("Error fetching todos:", error);
