@@ -1,10 +1,10 @@
 import axios from "axios";
-import { httpStatusCode } from "./http-status";
 
 // local vue api axios instance
 function localAxios() {
   const instance = axios.create({
-    baseURL: "http://i10c210.p.ssafy.io:8080", //스프링 API 호출 포트
+    // baseURL: "https://i10c210.p.ssafy.io", // run build
+    baseURL: "http://i10c210.p.ssafy.io:8080", // run dev
   });
 
   const token = localStorage.getItem('accessToken');
