@@ -33,7 +33,7 @@
         <div v-for="recordedVoice in recordedVoices" :key="recordedVoice .id" class="voice">
           <dt @click="openModal('RecordDetail', recordedVoice)">{{ recordedVoice.name }}</dt>
           <dd>
-            <button class="toggle-button" :class="{ 'active': recordedVoice .isActive }" @click="toggleVoice(recordedVoice )">
+            <button class="toggle-button" :class="{ 'active': recordedVoice .isActive }" @click="toggleVoice(recordedVoice)">
               {{ recordedVoice .isActive ? '사용 중' : '적용하기' }}
             </button>
             <button class="delete-button" @click="fnDelete(recordedVoice.id)">삭제</button>
@@ -49,6 +49,7 @@
 import { selectVoice, fetchVoiceList, fetchVoiceDetail, deleteVoice, } from '@/api/records'
 import RecordDetail from '@/components/Record/RecordDetail.vue'
 import AddRecord from '@/components/Record/AddRecord.vue'
+import { startLearning } from '@/api/learning'
 
 export default {
   name: 'Record',
@@ -76,22 +77,32 @@ export default {
         console.error(error)
       },
     async toggleVoice(voice) {
-      if (!voice.isActive) {
-        voice.isActive = true;
-        this.recordedVoices.forEach(v => {
-          if (v.id !== voice.id) v.isActive = false;
-          console.log('voice.id: ', voice.id)
-          console.log('v.id: ', v.id)
-        });
-
         try {
           // Use voice.id instead of recordId
           await selectVoice(voice.id, this.onSuccess, this.onFail);
         } catch (error) {
           console.error('Error selecting voice:', error);
         }
-      }
     },
+    
+    // async toggleVoice(voice) {
+    //   if (!voice.isActive) {
+    //     voice.isActive = true;
+    //     this.recordedVoices.forEach(v => {
+    //       if (v.id !== voice.id) v.isActive = false;
+    //       console.log('voice.id: ', voice.id)
+    //       console.log('v.id: ', v.id)
+    //     });
+
+    //     try {
+    //       // Use voice.id instead of recordId
+    //       await selectVoice(voice.id, this.onSuccess, this.onFail);
+    //     } catch (error) {
+    //       console.error('Error selecting voice:', error);
+    //     }
+    //   }
+    // },
+
     async openModal(component = 'RecordDetail', itemData = null) {
       if (component === 'RecordDetail' && itemData) {
         try {
@@ -224,7 +235,6 @@ export default {
   margin-left: 10px; /* 삭제 버튼 간격 조절 가능 */
   color: white;
 }
-
 .add-button {
   font-size: 20px;
   background-color: #ffffff;
